ml_local/requirements.txt
# PyTorch with CUDA support (REQUIRED for GPU training)
# Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# Or for CUDA 11.8: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# Or for CUDA 12.1: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# Check your CUDA version with: nvidia-smi
torch>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
tokenizers>=0.13.0
datasets>=2.14.0
supabase>=2.0.0
python-dotenv>=1.0.0
transformers>=4.30.0

